requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
anthropic>=0.7.0
requests>=2.31.0
langbaseai>=0.1.0  # Optional: for Langbase integration
